chore(migration): add admin seed in npm run start of backend
