+++ backend/package.json
@@ -45,7 +45,7 @@
     "dev": "nodemon --exec 'node --inspect=0.0.0.0 --require ts-node/register src/index.ts'",
     "dev2": "nodemon -x ts-node src/index.ts",
     "compile": "shx rm -rf lib/ && tsc",
-    "start": "knex migrate:latest && node lib/",
+    "start": "knex migrate:latest &&  knex seed:run --specific 00_admin_user.ts && node lib/",
     "prettier": "npx prettier \"**/*.ts\" --write",
     "mocha": "cross-env NODE_ENV=test mocha test/ --require ts-node/register --recursive --extension .ts --exit",
     "test": "xo && cross-env NODE_ENV=test npm run migrate && npm run mocha",
