+++ backend/src/hooks/user-registration.ts
@@ -3,7 +3,7 @@ import type { HookContext } from '../declarations'
 
 export const userRegistration = async (context: HookContext) => {
   // case of self registration or missing role property
-  if (context.params.provider && (!context.params.authentication || !context.data.role)) {
+  if (context.params.provider !== undefined && (!context.params.authentication || !context.data.role)) {
     context.data.role = 'guest'
   }
 }
