+++ docker-compose.yml
@@ -88,12 +88,12 @@ services:
       - CORS_ENABLED=${CORS_ENABLED} 
       - ROOT_PATH=
     expose:
-      - 5050
+      - 5053
     labels:
       - "traefik.http.routers.uploadFASTAPI.priority=1001"
       - "traefik.http.routers.uploadFASTAPI.tls=true"
       - "traefik.http.routers.uploadFASTAPI.rule=PathPrefix(`/upload-api`)"
-      - "traefik.http.services.uploadFASTAPI.loadbalancer.server.port=5050"
+      - "traefik.http.services.uploadFASTAPI.loadbalancer.server.port=5053"
       - "traefik.http.routers.uploadFASTAPI.middlewares=uploadFASTAPI-middleware"
       - "traefik.http.middlewares.uploadFASTAPI-middleware.stripprefix.prefixes=/upload-api"
     restart: unless-stopped
